Moved MultiSemanticLogger declaration, methods from semantic_logger.go to multi_semantic_logger.go
